agent/requirements.txt
langchain==0.1.0
wikipedia==1.4.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-anthropic==0.0.1
python-dotenv==1.0.0
pydantic==2.5.2
duckduckgo-search==4.1.1
langchain_google_genai==0.0.5
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
requests==2.31.0
beautifulsoup4==4.12.2